Shove aws runners into ~a locker~ the aws package
