+++ github-runner-provisioner/main.go
@@ -4,7 +4,6 @@ import (
 	"net/http"
 	"os"
 
-	"github.com/datawire/infra-actions/github-runner-provisioner/internal/aws"
 	"github.com/datawire/infra-actions/github-runner-provisioner/internal/monitoring"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 	log "github.com/sirupsen/logrus"
@@ -22,12 +21,10 @@ func init() {
 	log.SetLevel(log.InfoLevel)
 }
 
-var ec2Client *aws.Ec2Client
 var cfg *Config
 
 func main() {
 	cfg = NewConfig()
-	ec2Client = aws.NewEc2Client()
 
 	makeHandler := func(name string) http.Handler {
 		mux := http.NewServeMux()
